# src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Deepfake Detection API
  version: 1.0.0
  description: |
    API para detección de deepfakes en imágenes y videos usando Facia AI.
    
    ## Flujo de uso
    
    1. **Solicitar URL de subida** - `POST /api/files/upload-url`
    2. **Subir archivo a S3** - Usar la URL presignada desde el frontend
    3. **Iniciar análisis** - `POST /api/analysis/start`
    4. **Consultar resultado** - `POST /api/analysis/result`
    5. **Verificar estado** (opcional) - `GET /api/analysis/status/:id`
    
    ## Tipos de archivo soportados
    
    - **Imágenes:** PNG, JPG, JPEG (máx 50MB)
    - **Videos:** MP4, MOV, WEBM (máx 500MB)
    
    ## Tiempos de procesamiento
    
    - Imágenes: 2-10 segundos
    - Videos: 5-30 segundos

  contact:
    name: API Support
    email: support@tudominio.com
  license:
    name: MIT
servers:
  - url: https://backend-0ecl.onrender.com
    description: Producción
  - url: http://localhost:4000
    description: Desarrollo Local
...
paths:
...
  /auth/token:
    post:
      tags:
        - Auth
      summary: Obtener token de Facia AI (proxy autenticado)
      description: |
        Proxy autenticado para obtener un token de acceso de Facia AI.
        Requiere JWT válido del usuario. Las credenciales de Facia se pasan en el body.
      operationId: getFaciaToken
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                client_id:
                  type: string
                  description: Client ID de Facia AI
                client_secret:
                  type: string
                  description: Client Secret de Facia AI
      responses:
        "200":
          description: Token de Facia AI obtenido
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  access_token:
                    type: string
        "401":
          description: Token JWT requerido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error al contactar Facia AI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  #####################################
  ## GET /historial
  #####################################
  /historial:
    get:
      tags:
        - Historial
      summary: Obtener historial de análisis del usuario
      description: |
        Retorna el historial paginado de análisis realizados por el usuario autenticado.
        Requiere JWT válido.
      operationId: getUserHistory
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Número de página (default 1)
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: limit
          in: query
          description: Registros por página (default 10)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          example: 10
      responses:
        "200":
          description: Historial obtenido correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistorialResponse"
        "401":
          description: Token requerido o inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  #####################################
  ## POST /facia/liveness
  #####################################
...
components:
  schemas:
...
    HistorialResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        data:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
                example: "6476a2b1c3d4e5f6a7b8c9d0"
              referenceId:
                type: string
                example: "29D027498EE8EBB"
              isDeepfake:
                type: boolean
                example: true
              confidence:
                type: string
                example: "87.32"
              verdict:
                type: string
                enum: [FAKE, REAL]
                example: "FAKE"
              fileUrl:
                type: string
                format: uri
                example: "https://deepfake-detector-files.s3.amazonaws.com/video.mp4"
              createdAt:
                type: string
                format: date-time
                example: "2025-11-19T04:01:45Z"

    #################################
    # Files Schemas
    #################################
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Mensaje de error legible
          example: "No se pudo procesar la solicitud"
        details:
          type: string
          nullable: true
          description: Detalles técnicos (solo en desarrollo)
          example: "S3 Access Denied: Invalid credentials"
        code:
          type: string
          nullable: true
          description: Código de error interno
          example: "ERR_S3_ACCESS_DENIED"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Token JWT obtenido en /auth/login. Enviar como: Authorization: Bearer <token>"